[REF] web: convert module dependencies tests to hoot
task 3705027
